+++ nodes/sros/prompt.go
@@ -1,50 +1,9 @@
 package sros
 
 import (
-	"context"
 	"fmt"
-	"os"
 	"regexp"
-	"strings"
-
-	"github.com/charmbracelet/log"
-	"github.com/srl-labs/containerlab/clab/exec"
 )
-
-func (n *sros) setCustomPrompt(tplData *srosTemplateData) {
-	// when CLAB_CUSTOM_PROMPT is set to false, we don't generate custom prompt
-	if strings.ToLower(os.Getenv("CLAB_CUSTOM_PROMPT")) == "false" {
-		return
-	}
-
-	tplData.EnableCustomPrompt = true
-
-	// get the current prompt
-	prompt, err := n.currentPrompt(context.Background())
-	if err != nil {
-		log.Errorf("failed to get current prompt: %v", err)
-		tplData.EnableCustomPrompt = false
-		return
-	}
-
-	// adding newline to the prompt for better visual separation
-	tplData.CustomPrompt = "\\n" + prompt
-}
-
-// currentPrompt returns the current prompt extracted from the environment.
-func (n *sros) currentPrompt(ctx context.Context) (string, error) {
-	cmd, _ := exec.NewExecCmdFromString(`tail /home/sros/cf3:/bootlog.txt`)
-
-	execResult, err := n.RunExec(ctx, cmd)
-	if err != nil {
-		return "", err
-	}
-
-	log.Debugf("fetching prompt for node %s. stdout: %s, stderr: %s", n.Cfg.ShortName,
-		execResult.GetStdOutString(), execResult.GetStdErrString())
-
-	return getPrompt(execResult.GetStdOutString())
-}
 
 // getPrompt returns the prompt value from a string blob containing the prompt.
 // The s is the output of the "environment show | grep -A 2 prompt" command.
