+++ nodes/sros/template.go
@@ -1,0 +1,51 @@
+package sros
+
+// srosTemplateData top level data struct.
+type srosTemplateData struct {
+	Name            string
+	TLSKey          string
+	TLSCert         string
+	TLSAnchor       string
+	Banner          string
+	IFaces          map[string]tplIFace
+	SSHPubKeysRSA   []string
+	SSHPubKeysECDSA []string
+	MgmtMTU         int
+	MgmtIPMTU       int
+	DNSServers      []string
+	// EnableGNMIUnixSockServices enables GNMI unix socket services
+	// for the node. This is needed for "23.10 <= ver < 24.3" versions
+	EnableGNMIUnixSockServices bool
+	// EnableCustomPrompt enables custom prompt with added newline
+	// before the prompt.
+	EnableCustomPrompt bool
+	CustomPrompt       string
+	// SNMPConfig is a string containing SNMP configuration
+	SNMPConfig string
+	// GRPCConfig is a string containing GRPC configuration
+	GRPCConfig string
+	// ACLConfig is a string containing ACL configuration
+	ACLConfig string
+	// NetconfConfig is a string containing Netconf server configuration
+	NetconfConfig string
+	// EDAConfig is a string containing EDA configuration
+	EDAConfig string
+	// OCServerConfig is a string containing OpenConfig server configuration
+	OCServerConfig string
+	// SystemConfig is a string containing System configuration
+	SystemConfig string
+	// LoggingConfig is a string containing Logging configuration
+	LoggingConfig string
+	// SSHConfig is a string containing SSH configuration
+	SSHConfig string
+	//PartialConfig
+	PartialConfig string
+}
+
+// tplIFace template interface struct.
+type tplIFace struct {
+	Slot       string
+	Port       string
+	BreakoutNo string
+	Mtu        int
+}
